feat(util): make fibonacci function more general
For some reason I originally wrote the function with the assumption
that the caller would pass the limit to the fibonacci function.
This is of course limiting the usefullness of the function for
absolutely no reason.
We can just return an infinite iterator and rely on the caller to
do with that whatever they want.

Signed-off-by: Linda Siemons <linda@catbrained.dev>