tests: fix some tests and make cargo-mutants happy
There were some missed mutants. In particular some mutants caused
infinite loops, which were not caught by cargo-mutants and considered
as timeouts.
There already were tests that check for such infinite loops, but the
way that cargo-mutants currently works is that it waits for all tests
to finish, even when one of the tests for a mutant has already failed.
So the timeout tests failed, but the other tests kept running because
they didn't have a timeout.

To make the tests faster and the cargo-mutants results more conclusive,
I added a debug assertion in the code under test, which catches the
mutants that caused infinite loops.

This commit also includes a cargo-nextest config, which matches the
default, except for the timeout setting. Tests are now marked as slow
earlier than the default, and they actually get terminated after running
for too long (which the default apparently doesn't do).

Signed-off-by: Linda Siemons <linda@catbrained.dev>